add: Models for TS and update header styles
