+++ package.json
@@ -14,6 +14,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.11.0",
+    "react-image": "^4.1.0",
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.18.0"
   },
