+++ src/components/HeaderItem/HeaderItem.module.css
@@ -1,0 +1,13 @@
+.link{
+	text-decoration: none;
+	color: var(--dark-color);
+	font-size: 20px;
+}
+
+.content{
+	background-color: #b23d3d;
+	width: 100%;
+
+	display: grid;
+	place-items: center;
+}
