+++ src/components/HeaderItem/index.ts
@@ -1,0 +1,1 @@
+export * from './HeaderItem.tsx'
