--- src/components/Menu/index.ts
@@ -1,1 +1,0 @@
-export * from './Menu.tsx'
