+++ src/models/Header.model.ts
@@ -1,0 +1,5 @@
+export interface IHeaderItems {
+  name: string;
+  url: string;
+  id: string | number;
+}
