+++ src/pages/Error/Error.tsx
@@ -1,0 +1,11 @@
+import { Link } from "react-router-dom";
+
+
+export const Error = () => {
+  return (
+    <div>
+      Error page
+      <Link to={'/'}>To home</Link>
+    </div>
+  )
+}
